# src/main/resources/openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: Movie Ticket Booking API
  description: 'Browse screenings, reserve seats, complete bookings, cancel bookings, and administer the
    local catalog. This document is the source of truth: request/response models and the controller interfaces
    are generated from it at build time.'
  version: v1
servers:
- url: http://localhost:8080
  description: Local development
...
paths:
...
  /api/v1/bookings:
    get:
      tags:
      - Customer bookings
      summary: List owned booking history using an opaque cursor
      operationId: listBookings
      parameters:
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/CursorLimit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CursorPageResponseBookingResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
      - basicAuth: []
    post:
      tags:
      - Customer bookings
      summary: Pay for a reservation and create a booking
      operationId: createBooking
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        $ref: '#/components/parameters/IdempotencyKey'
        schema:
          type: string
          maxLength: 120
          minLength: 0
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequest'
            example:
              reservationId: 00000000-0000-4000-8000-000000000008
              paymentToken: tok_success
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessViolation'
      security:
      - basicAuth: []
  /api/v1/bookings/{id}/cancellations:
    post:
      tags:
      - Customer bookings
      summary: Cancel an owned booking idempotently
      operationId: cancelBooking
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: Idempotency-Key
        in: header
        required: true
        $ref: '#/components/parameters/IdempotencyKey'
        schema:
          type: string
          maxLength: 120
          minLength: 0
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CancellationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessViolation'
      security:
      - basicAuth: []
...
  /api/v1/bookings/{id}:
    get:
      tags:
      - Customer bookings
      summary: Get an owned booking
      operationId: getBooking
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
      - basicAuth: []
...
components:
  schemas:
...
    CreateBookingRequest:
      type: object
      properties:
        reservationId:
          type: string
          format: uuid
        discountCode:
          type: string
          maxLength: 40
          minLength: 0
        paymentToken:
          type: string
          maxLength: 100
          minLength: 1
      required:
      - paymentToken
      - reservationId
    BookingResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        reference:
          type: string
        screeningId:
          type: string
          format: uuid
        state:
          $ref: '#/components/schemas/BookingState'
        items:
          type: array
          items:
            $ref: '#/components/schemas/ItemResponse'
        subtotal:
          type: number
        discountAmount:
          type: number
        totalAmount:
          type: number
        currency:
          type: string
        paymentStatus:
          $ref: '#/components/schemas/PaymentStatus'
        refunds:
          type: array
          items:
            $ref: '#/components/schemas/RefundSummary'
        createdAt:
          type: string
          format: date-time
    ItemResponse:
      type: object
      properties:
        screeningSeatId:
          type: string
          format: uuid
        rowLabel:
          type: string
        seatNumber:
          type: integer
          format: int32
        category:
          $ref: '#/components/schemas/SeatCategory'
        unitPrice:
          type: number
    RefundSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        reason:
          $ref: '#/components/schemas/RefundReason'
        status:
          $ref: '#/components/schemas/RefundStatus'
        amount:
          type: number
    CancellationResponse:
      type: object
      properties:
        bookingId:
          type: string
          format: uuid
        bookingState:
          $ref: '#/components/schemas/BookingState'
        refund:
          $ref: '#/components/schemas/RefundResponse'
    RefundResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        bookingId:
          type: string
          format: uuid
        reason:
          $ref: '#/components/schemas/RefundReason'
        status:
          $ref: '#/components/schemas/RefundStatus'
        amount:
          type: number
        currency:
          type: string
...
    CursorPageResponseBookingResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/BookingResponse'
        nextCursor:
          type: string
...
    FieldViolation:
      type: object
      properties:
        field:
          type: string
          example: screeningSeatIds[1]
        code:
          type: string
          example: NotNull
        message:
          type: string
          example: The field is required.
      required:
      - code
      - field
      - message
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          format: uri
          example: /problems/seat-unavailable
        title:
          type: string
          example: Seat unavailable
        status:
          type: integer
          format: int32
          example: 409
        detail:
          type: string
          example: One or more requested seats are no longer available.
        instance:
          type: string
          format: uri
          example: /api/v1/seat-reservations
        code:
          type: string
          example: SEAT_UNAVAILABLE
        requestId:
          type: string
          format: uuid
          example: 8f3d8cbe-0e42-4d9b-9d50-96199b173ba7
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldViolation'
      required:
      - code
      - detail
      - instance
      - requestId
      - status
      - title
      - type
    SeatCategory:
      type: string
      enum:
      - REGULAR
      - PREMIUM
...
    BookingState:
      type: string
      enum:
      - PENDING_PAYMENT
      - CONFIRMED
      - FAILED
      - CANCELLED
    PaymentStatus:
      type: string
      enum:
      - INITIATED
      - SUCCEEDED
      - DECLINED
    RefundReason:
      type: string
      enum:
      - CANCELLATION
      - LATE_PAYMENT
      - SHOW_CANCELLED
    RefundStatus:
      type: string
      enum:
      - PENDING
      - PROCESSING
      - SUCCEEDED
      - FAILED
...
  responses:
    BadRequest:
      description: The request is malformed or fails validation.
      headers:
        X-Request-Id:
          description: Server-generated request correlation identifier
          schema:
            type: string
            format: uuid
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            title: Invalid request
            requestId: 8f3d8cbe-0e42-4d9b-9d50-96199b173ba7
            type: /problems/invalid-request
            status: 400
            instance: /api/v1/seat-reservations
            code: INVALID_REQUEST
            detail: The request could not be processed.
    Unauthenticated:
      description: Valid HTTP Basic credentials are required.
      headers:
        X-Request-Id:
          description: Server-generated request correlation identifier
          schema:
            type: string
            format: uuid
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            title: Authentication required
            requestId: 8f3d8cbe-0e42-4d9b-9d50-96199b173ba7
            type: /problems/authentication-required
            status: 401
            instance: /api/v1/seat-reservations
            code: AUTHENTICATION_REQUIRED
            detail: The request could not be processed.
    Forbidden:
      description: The authenticated account lacks the required role.
      headers:
        X-Request-Id:
          description: Server-generated request correlation identifier
          schema:
            type: string
            format: uuid
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            title: Access denied
            requestId: 8f3d8cbe-0e42-4d9b-9d50-96199b173ba7
            type: /problems/access-denied
            status: 403
            instance: /api/v1/seat-reservations
            code: ACCESS_DENIED
            detail: The request could not be processed.
    NotFound:
      description: The resource does not exist or is concealed from this account.
      headers:
        X-Request-Id:
          description: Server-generated request correlation identifier
          schema:
            type: string
            format: uuid
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            title: Resource not found
            requestId: 8f3d8cbe-0e42-4d9b-9d50-96199b173ba7
            type: /problems/resource-not-found
            status: 404
            instance: /api/v1/seat-reservations
            code: RESOURCE_NOT_FOUND
            detail: The request could not be processed.
    Conflict:
      description: The request conflicts with the current resource state.
      headers:
        X-Request-Id:
          description: Server-generated request correlation identifier
          schema:
            type: string
            format: uuid
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            title: Request conflict
            requestId: 8f3d8cbe-0e42-4d9b-9d50-96199b173ba7
            type: /problems/request-conflict
            status: 409
            instance: /api/v1/seat-reservations
            code: REQUEST_CONFLICT
            detail: The request could not be processed.
    BusinessViolation:
      description: A business rule prevents the operation.
      headers:
        X-Request-Id:
          description: Server-generated request correlation identifier
          schema:
            type: string
            format: uuid
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            title: Business rule violation
            requestId: 8f3d8cbe-0e42-4d9b-9d50-96199b173ba7
            type: /problems/business-rule-violation
            status: 422
            instance: /api/v1/seat-reservations
            code: BUSINESS_RULE_VIOLATION
            detail: The request could not be processed.
    InternalError:
      description: An unexpected failure occurred.
      headers:
        X-Request-Id:
          description: Server-generated request correlation identifier
          schema:
            type: string
            format: uuid
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            title: Internal server error
            requestId: 8f3d8cbe-0e42-4d9b-9d50-96199b173ba7
            type: /problems/internal-error
            status: 500
            instance: /api/v1/seat-reservations
            code: INTERNAL_ERROR
            detail: The request could not be processed.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: Unique key used to safely retry a state-changing request
      required: true
      schema:
        type: string
        maxLength: 120
        minLength: 1
      example: demo-reservation-001
...
    Cursor:
      name: cursor
      in: query
      description: Opaque cursor returned by the previous response
      schema:
        type: string
      example: eyJ2IjoxLCJrIjo...
    CursorLimit:
      name: limit
      in: query
      description: Number of records from 1 to 100
      schema:
        type: integer
        format: int32
        default: 20
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic